fix(project-details): fix focus rings and consolidate sort component
- Add focus-visible ring to accordion triggers in the About this
  project sidebar (section.tsx) and the Risks list group headers
  (listViewGroup.tsx), which previously showed the browser's default
  blue outline.
- Add focus-visible outline to the All invoice and All Timesheets
  links in the Tracking tab (costBurn.tsx, invoiceBurn.tsx).
- Replace the Risks tab's SortButton with the shared SortSelector
  design-system component (sortButton.tsx), so it gets the same
  keyboard/focus styling as the Projects list sort control.
- Remove the now-unused SortButton component from the design system
  and move its SortField/SortState/SortOrder types into SortSelector
  (sort-button/index.tsx deleted, sort-selector/index.tsx, index.ts).
