+++ frontend/packages/app/src/pages/project-details/about/section.tsx
@@ -26,7 +26,7 @@ export function Section({
       className="border-t border-outline-gray-1 first:border-t-0"
     >
       <Accordion.Header className="w-full flex justify-between items-center px-5 py-3">
-        <Accordion.Trigger className="w-full flex items-center gap-3 group flex-1">
+        <Accordion.Trigger className="w-full flex items-center gap-3 group flex-1 rounded-sm focus:outline-none focus-visible:ring-2 focus-visible:ring-outline-gray-3">
           <SmallDown
             aria-hidden
             className="size-4 shrink-0 text-ink-gray-7 transition-transform -rotate-90 group-data-panel-open:rotate-0"
