+++ frontend/packages/app/src/pages/project-details/tabs/risks/list/listViewGroup.tsx
@@ -33,7 +33,7 @@ export function RiskGroup({ value, label, risks }: RiskGroupProps) {
     <Accordion.Item value={value}>
       {/* Group header */}
       <Accordion.Header className="w-full">
-        <Accordion.Trigger className="flex items-center gap-2 w-full px-2 py-3 mb-2 text-base font-semibold text-ink-gray-8 border-b border-outline-gray-1 group">
+        <Accordion.Trigger className="flex items-center gap-2 w-full px-2 py-3 mb-2 text-base font-semibold text-ink-gray-8 border-b border-outline-gray-1 group rounded-sm focus:outline-none focus-visible:ring-2 focus-visible:ring-outline-gray-3 focus-visible:ring-inset">
           <SmallDown
             aria-hidden
             className="size-4 shrink-0 text-ink-gray-5 transition-transform -rotate-90 group-data-panel-open:rotate-0"
